check that we have valid data

+++ mod/post.php
@@ -598,18 +598,24 @@ function post_post(&$a) {
 			$ret['success'] = true;
 			$ret['pickup'] = array();
 			foreach($r as $rr) {
-				$x = json_decode($rr['outq_msg'],true);
+				if($rr['outq_msg']) {
+					$x = json_decode($rr['outq_msg'],true);
 
-				if(array_key_exists('message_list',$x)) {
-					foreach($x['message_list'] as $xx)
-						$ret['pickup'][] = array('notify' => json_decode($rr['outq_notify'],true),'message' => $xx);
+					if(! $x)
+						continue;
+
+					if(array_key_exists('message_list',$x)) {
+						foreach($x['message_list'] as $xx) {
+							$ret['pickup'][] = array('notify' => json_decode($rr['outq_notify'],true),'message' => $xx);
+						}
+					}
+					else
+						$ret['pickup'][] = array('notify' => json_decode($rr['outq_notify'],true),'message' => $x);
+
+					$x = q("delete from outq where outq_hash = '%s' limit 1",
+						dbesc($rr['outq_hash'])
+					);
 				}
-				else
-					$ret['pickup'][] = array('notify' => json_decode($rr['outq_notify'],true),'message' => $x);
-
-				$x = q("delete from outq where outq_hash = '%s' limit 1",
-					dbesc($rr['outq_hash'])
-				);
 			}
 		}
